@@ -7,20 +7,20 @@ authors = [{ name = 'WGBH-MLA', email = 'ryan_harbert@wgbh.org' }]
 readme = 'README.md'
 license = { text = 'MIT' }
 dependencies = [
-    "Django~=4.2",
-    "wagtail~=5.2",
-    "wagtail-factories~=4.1",
-    "pydantic~=2.6",
-    "psycopg2~=2.9",
-    "python-dotenv~=1.0",
-    "gunicorn~=21.2",
-    "wagtail-headless-preview~=0.7",
-    "django-cors-headers~=4.3",
-    "django-allauth~=0.60",
-    "elasticsearch~=8.12",
-    "wagtail-footnotes>=0.10.0",
+    'Django~=5.0',
+    'wagtail~=6.0',
+    'wagtail-factories~=4.1',
+    'pydantic~=2.6',
+    'psycopg2~=2.9',
+    'python-dotenv~=1.0',
+    'gunicorn~=21.2',
+    'wagtail-headless-preview~=0.7',
+    'django-cors-headers~=4.3',
+    'django-allauth~=0.60',
+    'elasticsearch~=8.12',
+    'wagtail-footnotes>=0.10.0',
 ]
-requires-python = '>=3.9,<4.0'
+requires-python = '>=3.10,<4.0'
 
 [project.urls]
 documentation = 'https://wgbh-mla.github.io/ov-wag/'
@@ -33,22 +33,18 @@ ov = 'cli:app'
 
 [project.optional-dependencies]
 cli = [
-    "typer~=0.9",
-    "loguru~=0.7",
-    "trogon~=0.5",
+    'typer~=0.9',
+    'loguru~=0.7',
+    'trogon~=0.5',
 ]
 production = [
     'gunicorn[standard]~=21.2',
     'django-storages[s3]~=1.14'
 ]
 
-dev = [
-    'black~=24.1',
-]
-
 [build-system]
-requires = ["pdm-backend"]
-build-backend = "pdm.backend"
+requires = ['pdm-backend']
+build-backend = 'pdm.backend'
 
 [tool.pdm]
 version = { source = 'file', path = 'ov_wag/_version.py' }
@@ -58,7 +54,8 @@ extend-exclude = 'migrations'
 
 [tool.ruff]
 extend-exclude = ['migrations']
-ignore = ['Q000']
+
+[lint]
 select = [
     'B', # flake8-bugbear
     'C4', # flake8-comprehensions
@@ -84,25 +81,26 @@ select = [
 
 [tool.pdm.dev-dependencies]
 dev = [
-    "ruff~=0.2",
-    "pre-commit~=3.6",
+    'black~=24.1',
+    'ruff~=0.2',
+    'pre-commit~=3.6',
 ]
 docs = [
-    "mkdocs~=1.5",
-    "mkdocs-material~=9.4",
-    "mike~=2.0",
-    "mkdocstrings[python]~=0.23",
-    "mkdocs-git-revision-date-localized-plugin~=1.2",
-    "mkdocs-jupyter~=0.24"
+    'mkdocs~=1.5',
+    'mkdocs-material~=9.4',
+    'mike~=2.0',
+    'mkdocstrings[python]~=0.23',
+    'mkdocs-git-revision-date-localized-plugin~=1.2',
+    'mkdocs-jupyter~=0.24'
 ]
 test = [
-    "pytest~=8.0",
-    "pytest-django~=4.8",
-    "pytest-cov~=4.1",
-    "pytest-sugar~=1.0",
-    "pytest-xdist~=3.5",
-    "nbmake~=1.5",
-    "coverage~=7.4",
+    'pytest~=8.0',
+    'pytest-django~=4.8',
+    'pytest-cov~=4.1',
+    'pytest-sugar~=1.0',
+    'pytest-xdist~=3.5',
+    'nbmake~=1.5',
+    'coverage~=7.4',
 ]
 
 [tool.pytest.ini_options]
